refactor(cli): split driver diagnostics tests
